figure sample conductivity

+++ S1 Superconductor/conductivity_example.py
@@ -8,7 +8,7 @@ def crit_cond(T, T_C):
     if T < T_C:
         return 0
     return float(T*T*T)
-crit_cond = np.vectorize(crit_cond, otypes=[np.float])
+crit_cond = np.vectorize(crit_cond, otypes=[float])
 
 T_C = 1
 T = np.linspace(0,2.5, num=10000)
